Admin functionality
Added a delete function to the admin panel.

Added sidebar table and editor.

Added details page for each page.

Added draggable order in the admin panel using jQuery UI, so you can reorder the pages as they will appear in the sites category list.

Added the ckedit package which I'll use in the future to add content to the body of the pages.

Co-Authored-By: yung lean <michael23b@users.noreply.github.com>